Rename composer to oscal2policy and fix name and description of CLI args
